requirements.txt
yt-dlp
customtkinter
Pillow
requests
